First attempt at project code.
Most things have been implemented, except for the stalling for the cycles of blinking of LED lights.

It seems not allowed to use timer functions. So no timer interrupts.

Also not yet tested the main function if it even does what it is supposed to without considering the blinking LED lights.