Tested code with simulator but stalling still not sure.
Tested the code in the simulator, should work as intended for the most part.

This means the inputs do actually give the proper outputs(for the most part).

Only main thing with simulator was unable to properly test for was timings of the stall/prolonged blinks. This means unsure if they LED turns on and off for the correct amount of time.